Moved the tests into tests/ folder
