+++ src/engine/src/timing.rs
@@ -64,49 +64,3 @@ impl Stopwatch {
         self.end_time = self.start_time;
     }
 }
-
-// Unit tests
-#[cfg(test)]
-mod test {
-    use super::*;
-    use std::thread;
-    use std::time;
-
-    #[test]
-    fn elapsed() {
-        let mut watch = Stopwatch::new();
-
-        watch.start();
-        thread::sleep(time::Duration::from_secs(2));
-        watch.stop();
-
-        assert_eq!(2, watch.elapsed().num_seconds());
-    }
-
-    #[test]
-    fn reset() {
-        let mut watch = Stopwatch::new();
-
-        watch.start();
-        thread::sleep(time::Duration::from_secs(2));
-        watch.stop();
-        watch.reset();
-
-        assert_eq!(0, watch.elapsed().num_nanoseconds().unwrap());
-    }
-
-    #[test]
-    fn restart() {
-        let mut watch = Stopwatch::new();
-
-        watch.start();
-        thread::sleep(time::Duration::from_secs(2));
-        watch.stop();
-
-        watch.restart();
-        thread::sleep(time::Duration::from_secs(1));
-        watch.stop();
-
-        assert_eq!(1, watch.elapsed().num_seconds());
-    }
-}
